cygwin.c: Fold two nearly identical functions into one
These functions are to convert from a windows path to posix, and vice
versa.  Except for the direction, the process is identical.  But the two
functions that go in a single direction have diverged slightly over
time.

The previous few commits were integrating the changes in one version
into the opposite one.

The two versions are now sufficiently alike that only direction-oriented
details differ, so the functions can be combined.

A new common function is created, and the other two functions merely
wrap it, calling it with a parameter to indicate the direction.