refcounted_he_(new|fetch)_pvn: Don't roll-own code
The new function utf8_to_bytes_new_pv() does a better job than these
code sections that are nearly duplicates of each other.

It's better for several reasons, such as that for long keys, it checks a
word at a time if it is downgradable.

This is a follow on to #22638, which was closed without the original
commits that this replaces.