regen/embed.pl: Handle m with p flags
Prior to this commit these were illegal.  This allows it when there is
no pesky thread context in the way.

This causes embed.fnc to generate macro 'Perl_foo' #defined to be  macro
'foo'.  This could be used to easily convert existing macros into having
long names should that become a name space pollution problem.

But more immediately, we have scattered around, various one line
functions that simply call something else.  Those were typically created
to preserve the pre-existing API should someone be using the long name,
and the implementation changed.

This commit allows those (that don't use thread context) to be
conveniently replaced by a macro.  The next couple commits will do that
for a couple of them.