Add utf8_hop_forward_overshoot()
This is like plain utf8_hop_forward() except it returns how many
characters the request would have overshot the edge if it had been
allowed to go beyond the edge.

This allows the caller to do error handling.

The code has to be changed to be more careful (than before this commit)
with counting the actual number of characters consumed in the hop.