ParseXS: add tests for RETVAL TARG optimisation
There is an optimisation whereby the OP_ENTERSUB's TARG is sometimes
used to return the result, rather than creating a new mortal.

This optimisation has been present for many years, but was untested.
This commit adds tests, including ones for custom typemap entries
of the sort that can still be optimised.

It also tests for the optimisation in conjunction with OUTLIST args.

It also adds tests for when the output code is specified explicitly
on the OUTPUT line rather than being retrieved from a typemap. Currently
this isn't actually optimised.