ParseXS: refactor: move RETVAL TARG optimise code
The perl code which does the emitting of the C code for the RETVAL value
return using TARG if possible, is in a different place than the general
RETVAL code emitting. This commit moves them into the same function.

For now this is a crude cut and paste: any simplifying of the resultant
function will be done later. For now, it's good just that more C code
generation functionality is concentrated in one place rather than being
diffused around all parts of ParseXS.pm.

In more detail: before this commit, the code looked like:

    sub process_file {
        ...
        if (RETVAL needs returning) {
            if (have template code from OUTPUT line) {
                # just directly emit the template code
            }
            elsif (can do TARG optimisation) {
                # directly emit code to do PUSHi() etc
            }
            else {
                $self->generate_output({ var => 'RETVAL', ...});
            }
        }
        ...
    }

    sub generate_output {
        ...
        if ($var eq 'RETVAL') {
                # emit generic C code to return RETVAL to the caller
        }
        ...
    }

After this commit, it looks like:

    sub process_file {
        ...
        if (RETVAL needs returning) {
            $self->generate_output({ var => 'RETVAL', ...});
        }
        ...
    }

    sub generate_output {
        ...
        if ($var eq 'RETVAL') {
            if (have template code from OUTPUT line) {
                # just directly emit the template code
                return;

            if (can do TARG optimisation) {
                # directly emit code to do PUSHi() etc
            }
            else {
                # emit generic C code to return RETVAL to the caller
            }
        }
        ...
    }