ParseXS: refactor: simplify dup OUTPUT RETVAL
If RETVAL appears twice in the OUTPUT section, it will be reported
as an error (using blurt()), but parsing will continue.
During parsing of the second RETVAL, it is treated as an ordinary
parameter rather than special-cased as RETVAL. This doesn't seem useful,
although ultimately it doesn't matter, as the parse will be reported as
failing at the end anyway, so it doesn't matter what code gets
generated in the meantime.

This commit makes RETVAL still be treated as RETVAL on second appearance
in OUTPUT, which get makes the code slightly simpler.

Also, change the dup-detecting lines above it from a postfix 'if' to a
block 'if', so that the flow-control 'next' appears prominently on its
own line.