ParseXS: add tests: 'RETVAL not in OUTPUT' warning
If the parser sees RETVAL in a CODE section, but there isn't an OUTPUT
section for the RETVAL, then it emits a warning. Test for this warning.

Note that it currently warns only if the OUTPUT section is not present;
it doesn't warn if the OUTPUT is present but doesn't contain RETVAL.
The next commit will fix this.