ParseXS: make RETVAL in CODE not OUTPUT warn more
ParseXS currently emits a warning if RETVAL is seen in a CODE section
but there is no OUTPUT section. This commit extends this so it warns even
if there *is* an OUTPUT section but that section doesn't contain
RETVAL.

To stop one of our existing tests giving a false positive warning, this
commit now makes it ignore 'PERL_UNUSED_VAR(RETVAL)' in a CODE section
in terms of 'have seen RETVAL in CODE', since this isn't a real use of
the variable.

Note that the original intent of adding this warning was that, without an

    OUTPUT
        RETVAL

the emitted code would still do XSRETURN(1), but nothing would be stored
at ST(0), so whatever was returned on the stack at that slot would be
whatever was already at that position on the stack, which might be one
of the passed args, or random garbage. This is true regardless of
whether OUTPUT is completely absent, or just RETVAL missing from OUTPUT.
So it's better to warn in both cases, not just the former.

Doing this also means that the 'xsub_seen_OUTPUT' field member is no longer
needed, simplifying the code.