ParseXS: refactor: make OUTPUT update param fields
Currently, an E::PXS::Node::Param object each created for each parameter
declared in the XSUB's signature. This object is then updated with any
extra data that appears in an INPUT line.

This commit extends that so that the Param object is also updated with
any extra data from an OUTPUT line.

This allows us to remove the $self->{xsub_map_varname_to_seen_in_OUTPUT}
field.

The next few commits will build upon this to simplify various bits of
code, now that the Param object contains OUTPUT state too.