ParseXS: refactor: make generate_output take Param
The generate_output() method currently takes a hash of args. The fields
of this hash are typically generated on the fly by each caller from the
fields of an ExtUtils::ParseXS::Node::Param object.

This commit makes it so that instead, that Param object is just passed
directly to generate_output(). It simplifies the code, and is a step
along the path to the eventual goal of making generate_output() be a
method of ExtUtils::ParseXS::Node::Param rather than of
ExtUtils::ParseXS.

This commit causes one slight change in behaviour: set magic; and this
is arguably a bug fix.

There are two ways of specifying args that get updated on return (the
XS equivalent of 'sub foo { $_[0]++ }'): either by declaring the parameter
as 'OUT' in the signature:

    int
    foo(OUT int a)

or by listing it on an OUTPUT line. For the latter, the arg has set
magic called on it after the update by default, but this can be
overridden with the SETMAGIC keyword:

    OUTPUT:
        arg1               // has a SETMAGIC() call
        SETMAGIC: DISABLE
        arg2               // doesn't have a SETMAGIC() call
        SETMAGIC: ENABLE
        arg3               // has a SETMAGIC() call

On the other hand, OUT params have set magic by default *unless*
there's an OUTPUT section with at least one SETMAGIC line, in which case
the last such SETMAGIC line dictates whether the OUT arg gets set magic.
This seems wrong, and this commit changes it so that OUT params *always*
get set magic.

If this turns out to be wrong, the line to look at is

    print "\tSvSETMAGIC($arg);\n" if !$param->{in_output} || $do_setmagic;

which says to do magic always for OUT params, and to honour the SETMAGIC
in force at the line for OUTPUT params.