ParseXS: fix recent OUTLIST breakage
The last couple of commits broke the unusual (and possibly not very sane)
combination of a parameter declared both as IN_OUTLIST and listed in an
OUTPUT section.

This means that generate_output() is called twice for the same
parameter, once to update the value of the arg's SV on the stack, and
once to push its value onto the stack.

So, distinguish between those two effects by whether the caller of
generate_output() passed an extra arg (for OUTLIST) or not, rather than
by whether the Param object has an in_out field of /OUTLIST$/.

This was manifesting itself as an extra PUSHs and SvSETMAGIC appearing
compared with 5.40.0.

Fix and add tests.