ParseXS: refactor: simplify RETVAL/TARG code
A recent commit moved, from outside to inside generate_output(), the
perl code which is responsible for generating the C code which handles
the "use TARG if possible when returning RETVAL" optimisation.

This commit now follows that up by simplifying and tidying that code.
Now that it's within generate_output(), the setting of $type, looking up
of the typemap etc, have all already been done at the start of the sub,
so there's no need to duplicate that effort.

Also, set {output_code} as a field value in the faked-up RETVAL
param object, so that inside generate_output() it can just check for
 a generic $param->{output_code} field rather than the special-cased
$self->{xsub_RETVAL_typemap_code} value.