ParseXS: refactor: move OUTPUT code handling.
OUTPUT sections can include explicit setting code to override the
normal typemap. For example:

    OUTPUT:
        arg1                              # use standard typemap entry
        arg2  my_setiv(ST(1), (IV)arg2);  # override

Formerly, the custom code was emitted directly during parsing by
OUTPUT_handler(). This commit moves that emitting code into
generate_output(), so that the emitting with/without override
are done in the same place.

This is a small part of an effort to concentrate all the code-emitting
parts into a small number of well-defined places.