ParseXS: add tests for RETVAL as a parameter
Normally RETVAL is an implicitly declared local var, but it is legal
to also declare it as an explicit parameter. Add tests for the various
permutations of this:

  - void and long XSUB return values;
  - no param type; int type in signature; int type in INPUT;
  - with and without an autocall.

This isn't well documented as to how it should be interpreted, so these
tests are more about checking current behaviour so that inadvertent
changes are detected, rather than approving the current behaviour.

Some of the current behaviour is less than satisfactory (generally
marked with 'XXX' in the test code comments).

In particular, a commit shortly is going to change how RETVAL is
implemented internally.

It also adds a few tests for NO_OUTPUT.