ParseXS: refactor: add RETVAL Node::Param object
Currently, RETVAL is handled almost entirely using special-case code
all over the place.

For non-void XSUBs, this commit adds a Node::Param object with a name of
RETVAL to the start of the list of Param objects in the Sig object - but
flagged as synthetic and no_init (so it gets declared but not
initialised, and doesn't consume an argument from the stack).

This makes the emitting of the

    some_type RETVAL;

declaration be part of general parameter processing, rather than being
special-cased. Soon, this new object will also be used to make emitting
of the code to return RETVAL be more regular; but for now just the
initial declaration is handled in the new fashion.

Note however, that things get complicated when RETVAL is also declared
as a parameter, e.g.

    int
    foo(RETVAL)
        long RETVAL
        ....

or

    int
    foo(long RETVAL)
        ....

or even as a parameter but without a type:

    int
    foo(RETVAL)
      # no INPUT lines
      ...

or even as an alien variable:

    int
    foo()
        long RETVAL = 99;

Much of the behaviour in such cases is unspecified, and is somewhat
emergent - i.e. it behaves that way due to the way the code happened to
be structured, rather than due to any particular intent. This commit
tries very hard to preserve the current behaviour - hence all the tests
added in the previous commit. It has however changed behaviour in a
couple of places where the old behaviour emitted broken C code. These
changes are indicated by the updates to t/001-basic.t in this commit.

To keep the backwards-compatibility behaviour, the basic approach of
this commit has been to make the RETVAL object be in one of three main
states:

fully-synthetic  What is initially pushed at the *start* of the param
                 list. RETVAL hasn't yet appeared in a signature or INPUT.
                 Its type is set to the XSUB's return type.

semi-real        Same as fully-synthetic, but with a defined
                 argument number, and with an updated position within
                 the parameter list. This signifies that a RETVAL has
                 appeared in the signature, but without a type yet
                 specified. It acts as a placeholder, causing one passed
                 argument to be skipped, and to appear in usage() error
                 messages etc.

real             The is_synthetic and no_init flags are turned off and
                 its type comes from the signature or INPUT line. This
                 is now just a normal parameter.