ParseXS: fix generate_output(): s/next/return/
A recent piece of refactoring by me, v5.41.4-92-g9c24c4ed29, moved
some error-checking code from OUTPUT_handler() to generate_output().
The former is in a loop, the latter isn't. So in

    if (bad) { blurt(..); next }

the 'next' needs changing to 'return'.

It turns out no tests yet exercise this error condition, so it it wasn't
spotted at the time.