ParseXS: refactor: use RETVAL param for output
The previous commit added a synthetic RETVAL Node::Param object and
used it for generating RETVAL declaration code rather than
special-casing it.

This commit extends this to use the object for output processing too,
removing more special-case handling, and removing the need for some
special state fields.

The next few commits will remove more RETVAL special-handling code and
generally simplify and clean up.

Note that this commit actually introduces a bug which will be fixed in
a few commits' time: the new code doesn't correctly handle multiple
CASEs, each with a RETVAL output, such as:

    int
    foo(...)
        CASE: x
            ...
            OUTPUT: RETVAL
        CASE: y
            ...
            OUTPUT: RETVAL

The code now sees this as a duplicate OUTPUT parameter. For now, a test
in t/002-more.t and t/XSMore.xs has been disabled until this is fixed.

(Actually, CASE with duplicate INPUT or OUTPUT sections has been broken
for a few weeks - but this is the first commit that actually caused a
test failure. All will be fixed soon...)