ParseXS: use XSUB return type for RETVAL return
Following on from the previous commit, which uses the new RETVAL
Node::Param object for generating output code, ensure that the
return type used for returning the XSUB's value is the XSUB's type
rather than RETVAL's type, where they differ.

For example, with

     int
     foo(long RETVAL)

RETVAL is is declared as long, but an int is returned to the caller.
Add tests to confirm.