ParseXS: refactor: better handle NO_OUTPUT
In an XSUB like

    NO_OUTPUT int
    foo(...)

An 'int RETVAL;' declaration is emitted, but no code is emitted to
return the value of RETVAL to the caller.

Before this commit, the way this was handled was that, after all INPUT
and CODE/PPCODE/autocall code had been emitted, ParseXS.pm did the
following:

    ($implicit_OUTPUT_RETVAL, $self->{xsub_return_type}) =
                                  (0, 'void') if $self->{xsub_seen_NO_OUTPUT};

That is to say, halfway through parsing the XSUB, it suddenly pretends
that the XSUB was declared void all along, and that there wasn't an
autocall (which would give an implicit OUTPUT: RETVAL declaration).

While this works, it is a confusing action-at-a-distance: at any
particular point in the code, you don't easily know whether

    if ($self->{xsub_return_type} eq 'void) { ... }

is dealing with a real void or fake void.

This commit removes that halfway-along state modification, and instead
handles NO_OUTPUT directly where needed:

For $implicit_OUTPUT_RETVAL, it now just never sets it if NO_OUTPUT is
present, rather than setting it and then later resetting it.

For $self->{xsub_return_type}, this now always represents the actual
declared return type of the XSUB, and whenever necessary,
$self->{xsub_seen_NO_OUTPUT} is checked along side
$self->{xsub_return_type} in various places.

In addition, the $var_num variable in OUTPUT_handler() handler has been
removed, as it was a vestigial value, now only used to determine whether
to call

    $self->blurt("Error: OUTPUT $outarg not a parameter")

which can be done directly as boolean expression.

Several new NO_OUTPUT tests have been added.