ParseXS: refactor: simplify ST() method.
The $self->ST($num) method typically emits "ST(" . ($num-1) . ")"
but it has special-case handling for $num being undef, or (in the case
of output), being 0.

The latter was a special-case for outputting RETVAL return code. Recent
refactoring means that the ST() method is no longer called with
$num == 0, so the method's code can be simplified (and the $is_output
parameter eliminated).

Should be no change in functionality.