ParseXS: fix CASE:
My recent merge commit v5.41.4-108-g9621dfa822 partially broke the CASE:
functionality; although I didn't spot it at first since the ParseXS test
suite only has a single CASE: test, and none of the XS code bundled with
perl core makes use of it.

The issue was that an XSUB with multiple CASEs has a single signature,
but potentially multiple sets of INPUT and OUTPUT sections. For example:

    int
    foo(abc, def)
        CASE: X
                int abc;
                int def;
            CODE:
                RETVAL = abc + def;
            OUTPUT:
                RETVAL

        CASE: Y
                long abc;
                long def;
            CODE:
                RETVAL = abc - def;
            OUTPUT:
                RETVAL

Once my merge commit made the data from INPUT and OUTPUT sections be
used to update each Node::Param object (which has been derived from the
signature entry) then this starts to fall apart. It only started
failing the single CASE test in t/002-more.t once *this* branch
made RETVAL an object (rather than it being special-cased).

The fix in this commit is a bit of hack. Once all of the XSUB's
signature has been parsed into a list of Node::Param objects, a copy of
this list is taken and kept as the original. Then each time the body is
parsed (typically once, but multiple times for bodies with CASE
keywords) that original list is copied and becomes the current parameter
list. The objects in that list are updated by INPUT and OUTPUT sections
as usual. Then for the start of the next CASE, the next copying blows
away all those updates.

If and when ParseXS takes on more of an AST-like appearance (with,
hypothetically, Node::INPUT, Node::OUTPUT etc objects) then this
approach should be revisited.

Note that this commit restores the CASE test in t/002-more.t which I
temporarily disabled a few commits ago.