ParseXS: fix under older perls
One line had: 'exists $param->{type}' as part of a condition. This
worked fine when $param was a normal hash ref (created under 'use
fields'), but the combination of:

- a recent commit which copies those hashes to $sig->{old_params}
  then resets some of the fields (to allow for multiple CASEs);

- older perls (in particular 5.8.9) which used pseudo-hashes to
  implement fields,

was causing the field to spring into existence, even though it wasn't defined.
So change the test from 'exists' to 'defined'.