ParseXS: refactor: use new() copying to simplify code
A recent commit needed, at one point, to make a copy of a Node::Param
object, and did so in a clunky way. But this can be simplified by
instead using the new() method's built-in ability to take an
initialisation hash ref as an arg. So this commit does that.

However, it turns out that this new() method *shared* rather than copied
that hash ref when running in an environment where 'use fields' isn't
available - namely during the building of perl itself. So fix that too.