utf8.h Reorder some bit flags
This changes the ordering of these bits (whose names are in the public
API, but whose values are perl's to set).  This is enables a future
commit which will replace a series of

if else if else ...

lines with a switch() statement, by placing the bit positions in the
order those existing clauses should be executed in.  The new ordering
will cause the switch() to preserve the current if clause execution
order.