ParseXS: add tests for typeless param placeholders
In XS, you can declare a parameter without a type being specified
(neither in the signature, nor in an INPUT line) and it informally
becomes a placeholder: no C variable of that name is declared, but it
causes the next arg to be skipped. For example,

    int
    foo(int A, B, int C)
        ...

emits C code which: checks for 3 args, declares A and C as vars, and
assigns them the values of ST(0) and ST(2).

This commit adds some tests which confirm this behaviour.

The semantics of such XSUBs with no body (i.e. autocall) are a bit odd.
The parameter is passed as an arg to the wrapped C function even though
the C var hasn't been declared or initialised. I decided to keep that
as-is, because it's possible that the var has instead been declared in a
PREINIT section. (I have no idea why you would want to do that, which
makes it almost certain that some XS module on CPAN is in fact doing
it.)

Note that placeholders with default values are supported: both the var
and the default value are ignored, and the arg is skipped. This is a bit
ugly, but it allows an XSUB to accept a variable number of parameters,
even if one of those parameters is no longer used.

(And like most such things, I first discovered that it was a "thing" by
seeing DB_File.xs break when I tried to disallow it. If there is a weird
or impossible thing to do in XS, you can be certain that DB_File is
doing it.)