ParseXS: allow SV* as placeholder
ExtUtils::ParseXS, up until and including perl 5.40.0, accepted
unparseable XSUB parameter declarations: if the parameter couldn't be
parsed, it was just silently skipped. This had the side effect that
any old line noise could be used as a placeholder: no parameter would
be declared, but during arg processing, the arg would be skipped.

During my recent refactoring work, I instead made it an error if the
parameter couldn't be parsed, but this broke some modules which were
using 'SV*' as a placeholder, e.g.

    int
    foo(int a, SV*, int b)
        ...

To partially restore backwards compatibility, this commit recognises
the special case of 'SV*' being used as a placeholder. All other
unparseable parameter declarations will continue to be an error.

It will however now still emit an error if 'SV*' is used on a bodiless sub.
It used to cause the call to the C function to be emitted as (from the
example above),

    RETVAL = foo(a, SV*, b);

which isn't legal C code. So making it an error isn't breaking backwards
compatibility.

The coder can still use C_ARGS, e.g.

    int
    foo(int a, SV*, int b)
        C_ARGS: a, b