utf8_to_uv_msgs: Revamp above Unicode code points handling
This is the most complicated of the problematic UTF-8 conditions.  There
are three types of these, each with a different warning message, and any
combination of them can be set to be warned or not warned about or an
error bit returned.  More rigorous testing, yet to be committed,  has
indicated that there are bugs in the existing implementation, which this
commit fixes.

The three cases are partially combined, so that if one case finds it is
not authorized to handle things, it drops to the next lower severity
one.