utf8_to_uv_msgs: Avoid extra deref pointer writes
This function is passed an address of where to write some values.
Instead of updating the derefernced pointer each time, do it once after
all the information is accumulated.

This allows for the values to be updated without updating the pointed to
variable, so makes some case in a switch() able to be more uniform.