builtin::trim: ensure TARG is properly reset
refactor a lot of custom "set SV to a string code" away to
sv_setpvn(), this:

 - fixed the original problem reported for #22784, where TARG wasn't
   being reset properly and contained a cached numeric version of the
   result from the previous call.

 - removed some never executed code, since builtin::trim is only XS
   and is not an OP with the TARGLEX optimization

 - fixes a possible problem if the result of the first call to trim()
   is COWed.

This does slightly change the taint behaviour, rather than making TARG
tainted iff source is tainted, it changes to the behaviour of the rest
of perl, making TARG tainted if any tainted input is seen in the
current expression.

See thr PR #22788 for some discussion on how we got here.

Fixes #22784