toke.c: remove PERL_CR_FILTER
When enabled (e.g. with `./Configure -A ccflags=-DPERL_CR_FILTER`), this
would define an implicit source filter that turns \r\n pairs into \n
whenever the parser reads a line of Perl source code.

As far as I can tell, perl already does the equivalent of this (without
a source filter) on all platforms and has done so since 5.005.

(Also, the code could read out-of-bounds if the last character of the
line read was a \r.)