Add utf8_to_uv_or_die()
This new function dies if the UTF-8 input to it is malformed.  There are
quite a few places in the core where we expect the input to be
wellformed, and just assume that it is.  This function is a drop-in
replacement for those, and we won't blindly continue if the assumption
is wrong.  There are also a bunch of places that don't make that
assumption, but check it and die immediately if malformed.  This
function replaces those too, along with the code needed to test the
return and die.